# services/graph_harness_api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HarnessArtifactListResponse": {
        "description": "List response for run artifacts.",
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/HarnessArtifactResponse"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "artifacts",
          "total"
        ],
        "title": "HarnessArtifactListResponse",
        "type": "object"
      },
      "HarnessArtifactResponse": {
        "description": "Serialized artifact payload.",
        "properties": {
          "content": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Content",
            "type": "object"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "key",
          "media_type",
          "content",
          "created_at",
          "updated_at"
        ],
        "title": "HarnessArtifactResponse",
        "type": "object"
      },
...
      "HarnessWorkspaceResponse": {
        "description": "Serialized workspace snapshot payload.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "snapshot": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Snapshot",
            "type": "object"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "snapshot",
          "created_at",
          "updated_at"
        ],
        "title": "HarnessWorkspaceResponse",
        "type": "object"
      },
...
      "JSONValue-Output": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "type": "object"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "type": "array"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MED13 Graph Harness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/spaces/{space_id}/runs/{run_id}/artifacts": {
      "get": {
        "description": "Return artifacts stored for one harness run.",
        "operationId": "list_artifacts_v1_spaces__space_id__runs__run_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessArtifactListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List artifacts for one harness run",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/v1/spaces/{space_id}/runs/{run_id}/artifacts/{artifact_key}": {
      "get": {
        "description": "Return one artifact stored for one harness run.",
        "operationId": "get_artifact_v1_spaces__space_id__runs__run_id__artifacts__artifact_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get one artifact for one harness run",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/runs/{run_id}/workspace": {
      "get": {
        "description": "Return the workspace snapshot stored for one harness run.",
        "operationId": "get_workspace_v1_spaces__space_id__runs__run_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessWorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get workspace snapshot for one harness run",
        "tags": [
          "artifacts"
        ]
      }
    },
...
  }
}